[#84980] ICPUWithExternalMmu: Take window ID in fault hook

+++ src/Emulator/Main/Peripherals/CPU/ICPUWithExternalMmu.cs
@@ -29,7 +29,7 @@ namespace Antmicro.Renode.Peripherals.CPU
 
         void SetMmuWindowPrivileges(ulong id, Privilege privileges);
 
-        void AddHookOnMmuFault(Action<ulong, AccessType, int> hook);
+        void AddHookOnMmuFault(Action<ulong, AccessType, ulong> hook);
 
         ulong GetMmuWindowStart(ulong id);
